except_wagtail/wsgi.py

"""
WSGI config for except_wagtail project.
It exposes the WSGI callable as a module-level variable named ``application``.
For more information on this file, see
https://docs.djangoproject.com/en/2.2/howto/deployment/wsgi/
"""

import os
import sys

from django.core.wsgi import get_wsgi_application

import dotenv


sys.path.append('/var/www/except.eco/except')

dotenv.read_dotenv(
    os.path.join(os.path.dirname(os.path.dirname(__file__)), '.env')
)
os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'except_wagtail.settings')

application = get_wsgi_application()
