requirements_current.txt
beautifulsoup4==4.6.0
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cssselect2==0.2.2
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.9
django-appconf==1.0.3
django-cogwheels==0.3
django-compressor==2.4
django-dotenv==1.4.2
django-libsass==0.8
django-modelcluster==5.0.1
django-modeltranslation==0.14.2
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
ffmpeg-python==0.2.0
future==0.18.2
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
libsass==0.19.4
mysqlclient==1.4.6
packaging==20.0
pathlib==1.0.1
pdf2image==1.11.0
Pillow==6.2.1
preview-generator==0.13
psycopg2==2.8.4
pycparser==2.19
pyexifinfo==0.4.0
pyparsing==2.4.6
PyPDF2==1.26.0
python-magic==0.4.15
pytz==2019.3
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.1.0
six==1.13.0
sqlparse==0.3.0
static3==0.7.0
tinycss2==1.0.2
Unidecode==1.1.1
urllib3==1.25.7
vtk==8.1.2
wagtail==2.7.1
wagtail-modeltranslation==0.10.6
wagtailmenus==3.0
Wand==0.5.8
webencodings==0.5.1
Willow==1.3
xvfbwrapper==0.2.9
